增加异常处理

+++ src/main/java/com/example/blog/blogcommunity/controller/PublishController.java
@@ -1,6 +1,5 @@
 package com.example.blog.blogcommunity.controller;
 
-import com.example.blog.blogcommunity.dto.QuestionDTO;
 import com.example.blog.blogcommunity.mapper.QuestionMapper;
 import com.example.blog.blogcommunity.model.Question;
 import com.example.blog.blogcommunity.model.User;
@@ -31,7 +30,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id,
+    public String edit(@PathVariable(name = "id")Long id,
                        Model model) {
         Question question = questionMapper.selectByPrimaryKey(id);
         model.addAttribute("title",question.getTitle());
@@ -48,7 +47,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request,
             Model model
